SFR-172 Basic Kinesis-to-s3 lambda for SFR/BookFinder
This is the core of a lambda that reads a Kinesis stream of Gutenberg
ePub urls and stores them in an s3 bucket. It reads the name and
the date_modified to determine if it should update the s3 records.

It uses the ETag as a md5 hash, which can be relied on as long as we
don't use multipart uploads or encrypt the files (won't encrypt, maybe
multipart?) Along with that it returns the rowID of for the item it
is part of along with the hash and a s3 url