SFR-163 Get and explode epubs records
This is a stream-only implementation of the .epub file fetching that stores
the resulting Buffer objects in s3 directly. This seems to perform well, but could
use review for other approaches if people have dealt with a similar situation.

The script stores the resulting files in two buckets and it has been verified that
epubjs can consume the resulting files from s3 to display epubs