SFR-193 Added Accessibility Scoring using Daisy/Ace
This generates a score and accessibility report using the Ace
library from Daisy. This contains basic metadata about the report
as well as the full JSON block produced for future reference.

The report takes some time to produce, though that may be to local
limitations, it should be reviewed for performance in the dev
environment.

This adds a new record type to the output Kinesis stream, as these
records should populate a new table in the db that relates to both
items (the score is calculated from the zipped version, but should
apply to both)