Update ePub parser to match new data model
Update the data input/output formats to match the new
model established for the persistence layer. This not only
improves reliability, it is necessary to include additional
identifiers to use in matching/merging returned data from this
function to the persistence layer.

The major change, in addition to adding standard data fields
to the output objects, is to include filenames as item record
identifiers. These provide unique identifiers for each source
and allow for better tracking of item-level metadata.