SFR-563 Improve Regex filename parsing/validation
This makes several improvements to how filenames are parsed by this function. First, it merges to separate regex steps into a single method in the `helpers` module. This is a better assignment of responsibilities and allows the function to utilize a single regex for both validation and parsing. Second, this makes the regex applied a more general pattern. Previously the check was limited to Project Gutenberg ePub files, which has been replaced with a regex that should generally validated `.epub` files while retaining the parsing necessary for Gutenberg's odd naming conventions.

Finally this adds some in-line documentation of the regex and the new method, which hopefully makes this more clear.

Additional tests have been updated and some functionality removed from `index.js` as this is now managed by the helper function.