SFR-563 Update Test Suite
An outdated test was confusingly setting a valid URL in a test designed to throw an error before overwriting it, which doesn't make sense as the URL will never be changed within the function. This simply sets the invalid URL for the test. The helper function's test file also did not include a test to raise the `regex-failure` error, and that has been added to account for all possible outputs of the function.