SFR-697 Add option to pass filename to function
This adds an option to explicitly pass a filename to use for the ePub file when storing/returing. This enables sources such as FrontiersIn to pass ePubs when the URLs do not contain the filenames (or anything that is usable as a filename. In this instance the name is extracted from a header field prior to passing to this function