SFR-697 Handle package.opf
Epub archives can have multiple structures. Previously this function was only checking for the presence of a `content.opf` file as an "entry-point" to the ebook. This is what is used by Project Gutenberg files, but other publishers use a `package.opf` file that serves the same purpose.

This update checks for either (no ePub file should have both) and uses it as the entry-point for the read-online version of the book.